requirements.txt
tensorflow-cpu
pandas == 1.0.4
numpy == 1.18.4
pathlib
streamlit == 0.62.0
nltk==3.5
spacy==2.3.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#egg=en_core_web_sm==2.3.1
